+++ packages/gamut-templates/src/GridForm/GridFormSubmit/index.tsx
@@ -1,8 +1,8 @@
-import { Button, Column } from '@codecademy/gamut';
+import { Button, Column } from '@codecademy/gamut-elements';
 import {
   ColumnSizes,
   ResponsiveProperty,
-} from '@codecademy/gamut/dist/Layout/types';
+} from '@codecademy/gamut-elements/dist/Layout/types';
 import React from 'react';
 
 export type GridFormSubmitProps = {