+++ packages/brand-components/src/Testimonial/__tests__/Testimonial-test.tsx
@@ -6,15 +6,15 @@ import { Testimonial } from '..';
 import styles from '../../styles.module.scss';
 
 const exampleTestmonial = {
-  name: 'Frank Reynolds',
+  firstName: 'Frank',
+  lastName: 'Reynolds',
   occupation: "Co-Owner @ Paddy's Pub",
-  imageUrl: '',
   quote:
     "I don't know how many years on this Earth I got left, I'm gonna get real weird with it.",
 };
 
 describe('Testimonial', () => {
-  it('adds the light class to the container name when its theme is light', () => {
+  it('adds the light class to the wrapper container when its theme is light', () => {
     const wrapper = mount(
       <Testimonial
         testimonial={exampleTestmonial}
@@ -24,14 +24,13 @@ describe('Testimonial', () => {
     );
 
     const containerClassName = wrapper
-      .find(`div`)
-      .first()
+      .find(`div.lightWrapper`)
       .prop('className');
 
     expect(containerClassName).toContain(styles.lightWrapper);
   });
 
-  it('adds the dark class to the container name when its theme is dark', () => {
+  it('adds the dark class to the wrapper container when its theme is dark', () => {
     const wrapper = mount(
       <Testimonial
         testimonial={exampleTestmonial}
@@ -41,10 +40,85 @@ describe('Testimonial', () => {
     );
 
     const containerClassName = wrapper
-      .find(`div`)
-      .first()
+      .find(`div.darkWrapper`)
       .prop('className');
 
     expect(containerClassName).toContain(styles.darkWrapper);
+  });
+
+  it('adds the small class to the content container when its size is small', () => {
+    const wrapper = mount(
+      <Testimonial
+        testimonial={exampleTestmonial}
+        size="small"
+        theme={VisualTheme.DarkMode}
+      />
+    );
+
+    const containerClassName = wrapper
+      .find('div.smallContainer')
+      .prop('className');
+
+    expect(containerClassName).toContain(styles.smallContainer);
+  });
+
+  it('adds the medium class to the content container when its size is medium', () => {
+    const wrapper = mount(
+      <Testimonial
+        testimonial={exampleTestmonial}
+        size="medium"
+        theme={VisualTheme.DarkMode}
+      />
+    );
+
+    const containerClassName = wrapper
+      .find('div.mediumContainer')
+      .prop('className');
+
+    expect(containerClassName).toContain(styles.mediumContainer);
+  });
+
+  it('adds the large class to the content container when its size is large', () => {
+    const wrapper = mount(
+      <Testimonial
+        testimonial={exampleTestmonial}
+        size="large"
+        theme={VisualTheme.DarkMode}
+      />
+    );
+
+    const containerClassName = wrapper
+      .find('div.largeContainer')
+      .prop('className');
+
+    expect(containerClassName).toContain(styles.largeContainer);
+  });
+
+  it('renders the Avatar component when an imageUrl prop is present', () => {
+    const wrapper = mount(
+      <Testimonial
+        testimonial={{ ...exampleTestmonial, imageUrl: 'someCoolUrl' }}
+        size="small"
+        theme={VisualTheme.DarkMode}
+      />
+    );
+
+    const avatarContainer = wrapper.find('div.avatarContainer');
+
+    expect(avatarContainer).toHaveLength(1);
+  });
+
+  it('does _not_ render the Avatar component when an imageUrl prop is _not_ present', () => {
+    const wrapper = mount(
+      <Testimonial
+        testimonial={exampleTestmonial}
+        size="small"
+        theme={VisualTheme.DarkMode}
+      />
+    );
+
+    const avatarContainer = wrapper.find('div.avatarContainer');
+
+    expect(avatarContainer).toHaveLength(0);
   });
 });