+++ packages/gamut/src/Tabs/TabPanel/index.tsx
@@ -1,4 +1,7 @@
 import React, { ReactNode, FunctionComponent } from 'react';
+import cx from 'classnames';
+
+import s from './styles.module.scss';
 
 export type TabPanelProps = {
   active?: boolean;
@@ -18,9 +21,12 @@ export const TabPanel: FunctionComponent<TabPanelProps> = ({
   <div
     id={id}
     aria-labelledby={id.replace('-panel', '')}
+    aria-hidden={!active}
     role="tabpanel"
-    style={{ display: active ? 'block' : 'none' }}
-    className={className}
+    className={cx(className, {
+      [s.active]: active,
+      [s.hidden]: !active,
+    })}
   >
     {active || renderAllPanels ? children : null}
   </div>